refactor: 불필요한 코드 제거 및 수정 (#21)
* 패키지명 outport -> service
* packageInfo 제거

3tier 에서 outport가 부적절해 보여 service로 수정했습니다.
