--- src/main/java/com/example/green/infra/package-info.java
@@ -1,1 +1,0 @@
-package com.example.green.infra;
