--- src/test/java/com/example/green/domain/TestEntityTest.java
@@ -1,11 +1,0 @@
-package com.example.green.domain;
-
-import org.junit.jupiter.api.Test;
-
-class TestEntityTest {
-
-	@Test
-	void 테스트_엔티티_의미없는_테스트() {
-		new TestEntity();
-	}
-}
