--- src/test/java/com/example/green/package-info.java
@@ -1,1 +1,0 @@
-package com.example.green;
